Hoist GuidMapperColumn into API
